+++ example/Test.hs
@@ -5,34 +5,43 @@
 import Prelude ()
 import Prelude.Compat
 import AuthAPI (app, authSettings, LoginForm(..), homePage, loginPage)
-import Control.Concurrent (threadDelay)
-import Control.Exception.Base (bracket)
-import Control.Monad (guard, when)
 import Network.Wai (Application)
 import Test.Hspec (Spec, hspec, describe, it)
-import Test.Hspec.Wai (matchBody, shouldRespondWith, liftIO, with, request, get, post)
+import Test.Hspec.Wai (shouldRespondWith, with, request, get)
 import Text.Blaze.Renderer.Utf8 (renderMarkup)
 import Servant (Proxy(..))
-import Servant.Server (serve)
-import System.IO (openTempFile, hClose)
 import qualified Data.ByteString.Lazy as BSL
-import qualified Data.ByteString.Lazy.Char8 as BSLC8
 import Crypto.Random (drgNew)
-import Servant (FormUrlEncoded(..), contentType)
+import Servant (FormUrlEncoded, contentType)
 import Servant.Server.Experimental.Auth.Cookie
 import Network.HTTP.Types (methodGet, methodPost, hContentType, hCookie)
 import Network.HTTP.Media.RenderHeader (renderHeader)
 import Network.Wai.Test (SResponse(..))
 
-#if MIN_VERSION_servant (0,9,0)
-import Web.FormUrlEncoded (toForm)
+#if MIN_VERSION_hspec_wai (0,7,0)
+import Test.Hspec.Wai.Matcher (bodyEquals, ResponseMatcher(..), MatchBody)
 #else
-import Servant (ToFormUrlEncoded, toFormUrlEncoded, mimeRender)
+import Test.Hspec.Wai (matchBody)
+#endif
+
+#if MIN_VERSION_servant (0,9,0)
+import Web.FormUrlEncoded (ToForm, toForm, urlEncodeForm)
+#else
+import Servant (ToFormUrlEncoded, mimeRender)
 #endif
 
 
+#if MIN_VERSION_hspec_wai (0,7,0)
+matchBody' :: BSL.ByteString -> MatchBody
+matchBody' = bodyEquals
+#else
+matchBody' :: BSL.ByteString -> Maybe BSL.ByteString
+matchBody' = Just
+#endif
+
 #if MIN_VERSION_servant (0,9,0)
--- TODO
+encode :: ToForm a => a -> BSL.ByteString
+encode = urlEncodeForm . toForm
 #else
 encode :: ToFormUrlEncoded a => a -> BSL.ByteString
 encode = mimeRender (Proxy :: Proxy FormUrlEncoded)
@@ -52,13 +61,13 @@ spec = with mkApp $ do
   describe "home page" $ do
     it "responds successfully" $ do
       get "/" `shouldRespondWith` 200 {
-        matchBody = Just . renderMarkup $ homePage
+        matchBody = matchBody' $ renderMarkup homePage
         }
 
   describe "login page" $ do
     it "responds successfully" $ do
       get "/login" `shouldRespondWith` 200 {
-        matchBody = Just . renderMarkup $ loginPage True
+        matchBody = matchBody' $ renderMarkup (loginPage True)
         }
 
     it "shows message on incorrect login" $ do
@@ -68,15 +77,13 @@ spec = with mkApp $ do
           }
       let r = request methodPost "/login" [formContentType] loginForm
       r `shouldRespondWith` 200 {
-        matchBody = Just . renderMarkup $ loginPage False
+        matchBody = matchBody' $ renderMarkup (loginPage False)
         }
-
 
   describe "private page" $ do
     it "rejects requests without cookies" $ do
       let r = get "/private"
-      r `shouldRespondWith` 403 { matchBody = Just "User doesn't exist" }
-
+      r `shouldRespondWith` 403 { matchBody = matchBody' "User doesn't exist" }
 
     it "accepts requests with proper cookies" $ do
       let loginForm = encode $ LoginForm {
@@ -89,6 +96,7 @@ spec = with mkApp $ do
 
       let r = request methodGet "/private" [(hCookie, cookieValue)] ""
       r `shouldRespondWith` 200
+
 
 mkApp :: IO Application
 mkApp = do